requirements.txt
psycopg2-binary
pandas>=2.2
python-dateutil>=2.9
